FIR filter controller
